+++ Realm/RLMSyncSession.mm
@@ -69,7 +69,20 @@ using namespace realm;
 
 @end
 
+@interface RLMSyncSession ()
+@property (class, nonatomic, readonly) dispatch_queue_t notificationsQueue;
+@end
+
 @implementation RLMSyncSession
+
++ (dispatch_queue_t)notificationsQueue {
+    static dispatch_queue_t queue;
+    static dispatch_once_t onceToken;
+    dispatch_once(&onceToken, ^{
+        queue = dispatch_queue_create("io.realm.sync.sessionsNotificationQueue", DISPATCH_QUEUE_SERIAL);
+    });
+    return queue;
+}
 
 - (instancetype)initWithSyncSession:(std::shared_ptr<SyncSession>)session {
     if (self = [super init]) {
@@ -153,17 +166,15 @@ using namespace realm;
         if (session->state() == SyncSession::PublicState::Error) {
             return nil;
         }
-        // Get the current runloop, or create one if necessary.
-        CFRunLoopRef currentRunLoop = CFRunLoopGetCurrent();
+        dispatch_queue_t queue = RLMSyncSession.notificationsQueue;
         auto notifier_direction = (direction == RLMSyncProgressDirectionUpload
                                    ? SyncSession::NotifierType::upload
                                    : SyncSession::NotifierType::download);
         bool is_streaming = (mode == RLMSyncProgressReportIndefinitely);
         uint64_t token = session->register_progress_notifier([=](uint64_t transferred, uint64_t transferrable) {
-            CFRunLoopPerformBlock(currentRunLoop, kCFRunLoopCommonModes, ^{
+            dispatch_async(queue, ^{
                 block((NSUInteger)transferred, (NSUInteger)transferrable);
             });
-            CFRunLoopWakeUp(currentRunLoop);
         }, notifier_direction, is_streaming);
         return [[RLMProgressNotificationToken alloc] initWithTokenValue:token session:std::move(session)];
     }