+++ RealmSwift/Sync.swift
@@ -582,12 +582,12 @@ public extension SyncSession {
     /**
      Register a progress notification block.
 
+     If the session has already received progress information from the
+     synchronization subsystem, the block will be called immediately. Otherwise, it
+     will be called as soon as progress information becomes available.
+
      Multiple blocks can be registered with the same session at once. Each block
-     will be invoked from the runloop of the thread on which it was registered,
-     creating a new runloop if none exists. If the session has already received
-     progress information from the synchronization subsystem, the block will be
-     called immediately. Otherwise, it will be called as soon as progress
-     information becomes available.
+     will be invoked on a side queue devoted to progress notifications.
 
      The token returned by this method must be retained as long as progress
      notifications are desired, and the `stop()` method should be called on it